requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
aiogram==3.4.1
httpx==0.27.0
pillow==10.3.0
pydantic==2.5.3
python-dotenv==1.0.1
loguru==0.7.2
